MDEV-27036: re-enable my_json_writer-t unit test
