+++ CMakeLists.txt
@@ -142,6 +142,7 @@ IF (NOT CPACK_GENERATOR)
   ENDIF(WIN32) 
 ENDIF(NOT CPACK_GENERATOR)
 
+INCLUDE(build_depends)
 INCLUDE(FeatureSummary)
 INCLUDE(misc)
 INCLUDE(mysql_version)
@@ -537,8 +538,6 @@ INSTALL_DOCUMENTATION(README.md CREDITS COPYING THIRDPARTY COMPONENT Readme)
 IF(UNIX)
   INSTALL_DOCUMENTATION(Docs/INSTALL-BINARY Docs/README-wsrep COMPONENT Readme)
 ENDIF()
-
-INCLUDE(build_depends)
 
 INCLUDE(CPack)
