+++ cmake/ctest.cmake
@@ -22,3 +22,4 @@ MACRO(MY_ADD_TESTS)
   ENDFOREACH()
 ENDMACRO()
 
+FIND_PACKAGE(Boost COMPONENTS unit_test_framework)
