+++ sql/set_var.h
@@ -247,8 +247,7 @@ protected:
     Typically it's the same as session_value_ptr(), but it's different,
     for example, for ENUM, that is printed as a string, but stored as a number.
   */
-  uchar *session_var_ptr(THD *thd) const
-  { return ((uchar*)&(thd->variables)) + offset; }
+  uchar *session_var_ptr(THD *thd) const;
 
   uchar *global_var_ptr() const
   { return ((uchar*)&global_system_variables) + offset; }
