+++ sql/sql_explain.h
@@ -54,6 +54,8 @@ it into the slow query log.
 #ifndef SQL_EXPLAIN_INCLUDED
 #define SQL_EXPLAIN_INCLUDED
 
+class select_result_sink;
+
 class String_list: public List<char>
 {
 public:
