+++ unittest/sql/CMakeLists.txt
@@ -32,5 +32,6 @@ MY_ADD_TEST(mf_iocache)
 
 # Json writer needs String which needs sql library
 ADD_EXECUTABLE(my_json_writer-t my_json_writer-t.cc dummy_builtins.cc)
-TARGET_LINK_LIBRARIES(my_json_writer-t sql mytap)
+TARGET_LINK_LIBRARIES(my_json_writer-t PUBLIC sql mytap)
+TARGET_COMPILE_DEFINITIONS(my_json_writer-t PUBLIC JSON_WRITER_UNIT_TEST)
 MY_ADD_TEST(my_json_writer)
